backend/requirements.txt
# Web Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0

# AI & ML
anthropic==0.7.8
openai==1.3.5
sentence-transformers==2.2.2
chromadb==0.4.18

# Database
sqlalchemy==2.0.23
alembic==1.12.1
psycopg2-binary==2.9.9

# Data Processing
pydantic==2.5.0
pandas==2.1.3
numpy==1.25.2

# Vector Search (choose one)
pinecone-client==2.2.4
# chromadb==0.4.18

# Security & Auth
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# Utilities
python-dotenv==1.0.0
httpx==0.25.2
aiofiles==23.2.1
python-pdf==4.3.1
reportlab==4.0.7

# Development
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.11.0
isort==5.12.0
mypy==1.7.1
pre-commit==3.6.0

# Monitoring
structlog==23.2.0
sentry-sdk[fastapi]==1.38.0